backend/ai_service/requirements.txt
# Core web framework
fastapi
uvicorn

# Azure OpenAI + LangChain
openai
langchain
langchain-community
langchain-openai
python-dotenv

# MongoDB integration
pymongo

# Typing and utility
pydantic
